Allow feedback from Gitlab instead of Jenkins
Build time estimation is missing
https://gitlab.com/gitlab-org/gitlab-ce/issues/13412

Test results are not exposed either
https://gitlab.com/gitlab-org/gitlab-ce/issues/17081

Some hacking around necessary to munge the Gitlab CI model into
something feedback-shaped - this is probably not the best way
to do it but it'll do for now.

jQuery doesn't like '/' in element ids (which to be fair is only
valid in html5) so again a hack in PublishableJob is enough for now.